requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
pytz>=2021.1
tabulate>=0.9.0
pandas>=1.5.0
jinja2>=3.0.0
